Refactored files to folders for more cohesive display
